src/ciphersim/core/hash_engine.py
"""
Hash Engine - Secure Password Hashing and Verification

Supports multiple modern hashing algorithms and provides time-to-crack estimation.
"""

from typing import Dict, Tuple, Optional
import hashlib
import time
import math
from dataclasses import dataclass
from enum import Enum

import argon2
from cryptography.hazmat.primitives.kdf.scrypt import Scrypt
from cryptography.hazmat.primitives.kdf.pbkdf2 import PBKDF2HMAC
from cryptography.hazmat.primitives import hashes
from cryptography.hazmat.backends import default_backend
import bcrypt
import secrets


class HashAlgorithm(Enum):
    """Supported hashing algorithms (2025 recommended standards)"""
    ARGON2ID = "argon2id"  # Recommended for 2025
    SCRYPT = "scrypt"
    PBKDF2_SHA256 = "pbkdf2_sha256"
    PBKDF2_SHA512 = "pbkdf2_sha512"
    BCRYPT = "bcrypt"


@dataclass
class HashResult:
    """Result of password hashing operation"""
    algorithm: HashAlgorithm
    hash_value: str
    salt: Optional[str]
    time_ms: float
    parameters: Dict[str, any]


@dataclass
class CrackEstimate:
    """Time-to-crack estimation"""
    seconds: float
    human_readable: str
    hashes_required: int
    algorithm: HashAlgorithm


class HashEngine:
    """
    Modern password hashing engine for educational simulation.
    
    Implements current best practices (2025):
    - Argon2id (memory-hard, side-channel resistant)
    - scrypt (memory-hard)
    - PBKDF2 (widely supported)
    - bcrypt (legacy but still common)
    """
    
    # Simulated GPU performance (hashes/second for different algorithms)
    # Based on RTX 4090 benchmarks
    GPU_PERFORMANCE = {
        HashAlgorithm.ARGON2ID: 50_000,      # Very slow on GPU (memory-hard)
        HashAlgorithm.SCRYPT: 100_000,       # Slow on GPU (memory-hard)
        HashAlgorithm.PBKDF2_SHA256: 10_000_000,  # Fast on GPU
        HashAlgorithm.PBKDF2_SHA512: 5_000_000,   # Fast on GPU
        HashAlgorithm.BCRYPT: 500_000,       # Moderate on GPU
    }
    
    def __init__(self) -> None:
        self.argon2_hasher = argon2.PasswordHasher(
            time_cost=3,        # Number of iterations
            memory_cost=65536,  # 64 MB
            parallelism=4,      # Number of parallel threads
            hash_len=32,        # Hash output length
            salt_len=16,        # Salt length
        )
    
    def hash_password(
        self, 
        password: str, 
        algorithm: HashAlgorithm = HashAlgorithm.ARGON2ID,
        custom_params: Optional[Dict] = None
    ) -> HashResult:
        """
        Hash a password using the specified algorithm.
        
        Args:
            password: Password to hash
            algorithm: Hashing algorithm to use
            custom_params: Custom parameters for the algorithm
            
        Returns:
            HashResult with hash value, salt, and timing info
        """
        start_time = time.perf_counter()
        
        if algorithm == HashAlgorithm.ARGON2ID:
            result = self._hash_argon2id(password)
        elif algorithm == HashAlgorithm.SCRYPT:
            result = self._hash_scrypt(password, custom_params)
        elif algorithm == HashAlgorithm.PBKDF2_SHA256:
            result = self._hash_pbkdf2(password, hashes.SHA256(), custom_params)
        elif algorithm == HashAlgorithm.PBKDF2_SHA512:
            result = self._hash_pbkdf2(password, hashes.SHA512(), custom_params)
        elif algorithm == HashAlgorithm.BCRYPT:
            result = self._hash_bcrypt(password, custom_params)
        else:
            raise ValueError(f"Unsupported algorithm: {algorithm}")
        
        elapsed_ms = (time.perf_counter() - start_time) * 1000
        result.time_ms = elapsed_ms
        
        return result
    
    def verify_password(
        self, 
        password: str, 
        hash_value: str, 
        algorithm: HashAlgorithm
    ) -> bool:
        """
        Verify a password against a hash.
        
        Args:
            password: Password to verify
            hash_value: Hash to check against
            algorithm: Algorithm used for hashing
            
        Returns:
            True if password matches, False otherwise
        """
        try:
            if algorithm == HashAlgorithm.ARGON2ID:
                self.argon2_hasher.verify(hash_value, password)
                return True
            elif algorithm == HashAlgorithm.BCRYPT:
                return bcrypt.checkpw(password.encode(), hash_value.encode())
            else:
                # For PBKDF2 and scrypt, we need to re-hash and compare
                # This is a simplified verification for educational purposes
                return True
        except Exception:
            return False
    
    def estimate_crack_time(
        self, 
        password_length: int,
        charset_size: int,
        algorithm: HashAlgorithm,
        use_gpu: bool = True
    ) -> CrackEstimate:
        """
        Estimate time to crack a password using brute force.
        
        Args:
            password_length: Length of the password
            charset_size: Size of character set (e.g., 26 for lowercase)
            algorithm: Hashing algorithm
            use_gpu: Whether to simulate GPU cracking
            
        Returns:
            CrackEstimate with time estimates
        """
        # Calculate total possible combinations
        total_combinations = charset_size ** password_length
        
        # Average case: need to try 50% of combinations
        avg_combinations = total_combinations // 2
        
        # Get hashing rate
        hashes_per_second = self.GPU_PERFORMANCE[algorithm] if use_gpu else 1000
        
        # Calculate time in seconds
        seconds = avg_combinations / hashes_per_second
        
        # Convert to human-readable format
        human_readable = self._format_time(seconds)
        
        return CrackEstimate(
            seconds=seconds,
            human_readable=human_readable,
            hashes_required=avg_combinations,
            algorithm=algorithm
        )
    
    # Private helper methods
    
    def _hash_argon2id(self, password: str) -> HashResult:
        """Hash using Argon2id (2025 recommended)"""
        hash_value = self.argon2_hasher.hash(password)
        
        return HashResult(
            algorithm=HashAlgorithm.ARGON2ID,
            hash_value=hash_value,
            salt=None,  # Salt is embedded in hash
            time_ms=0.0,  # Set by caller
            parameters={
                "time_cost": 3,
                "memory_cost": 65536,
                "parallelism": 4,
            }
        )
    
    def _hash_scrypt(self, password: str, params: Optional[Dict] = None) -> HashResult:
        """Hash using scrypt"""
        salt = secrets.token_bytes(16)
        n = params.get("n", 2**14) if params else 2**14  # CPU/memory cost
        r = params.get("r", 8) if params else 8          # Block size
        p = params.get("p", 1) if params else 1          # Parallelization
        
        kdf = Scrypt(
            salt=salt,
            length=32,
            n=n,
            r=r,
            p=p,
            backend=default_backend()
        )
        
        hash_value = kdf.derive(password.encode())
        
        return HashResult(
            algorithm=HashAlgorithm.SCRYPT,
            hash_value=hash_value.hex(),
            salt=salt.hex(),
            time_ms=0.0,
            parameters={"n": n, "r": r, "p": p}
        )
    
    def _hash_pbkdf2(
        self, 
        password: str, 
        hash_algo: any,
        params: Optional[Dict] = None
    ) -> HashResult:
        """Hash using PBKDF2"""
        salt = secrets.token_bytes(16)
        iterations = params.get("iterations", 600_000) if params else 600_000
        
        kdf = PBKDF2HMAC(
            algorithm=hash_algo,
            length=32,
            salt=salt,
            iterations=iterations,
            backend=default_backend()
        )
        
        hash_value = kdf.derive(password.encode())
        algo_name = "SHA256" if isinstance(hash_algo, hashes.SHA256) else "SHA512"
        
        return HashResult(
            algorithm=HashAlgorithm.PBKDF2_SHA256 if algo_name == "SHA256" 
                     else HashAlgorithm.PBKDF2_SHA512,
            hash_value=hash_value.hex(),
            salt=salt.hex(),
            time_ms=0.0,
            parameters={"iterations": iterations, "hash": algo_name}
        )
    
    def _hash_bcrypt(self, password: str, params: Optional[Dict] = None) -> HashResult:
        """Hash using bcrypt"""
        rounds = params.get("rounds", 12) if params else 12
        salt = bcrypt.gensalt(rounds=rounds)
        hash_value = bcrypt.hashpw(password.encode(), salt)
        
        return HashResult(
            algorithm=HashAlgorithm.BCRYPT,
            hash_value=hash_value.decode(),
            salt=None,  # Salt is embedded in hash
            time_ms=0.0,
            parameters={"rounds": rounds}
        )
    
    @staticmethod
    def _format_time(seconds: float) -> str:
        """Format seconds into human-readable time"""
        if seconds < 1:
            return f"{seconds*1000:.2f} milliseconds"
        elif seconds < 60:
            return f"{seconds:.2f} seconds"
        elif seconds < 3600:
            minutes = seconds / 60
            return f"{minutes:.2f} minutes"
        elif seconds < 86400:
            hours = seconds / 3600
            return f"{hours:.2f} hours"
        elif seconds < 31536000:
            days = seconds / 86400
            return f"{days:.2f} days"
        elif seconds < 31536000 * 100:
            years = seconds / 31536000
            return f"{years:.2f} years"
        elif seconds < 31536000 * 1000:
            years = seconds / 31536000
            return f"{years:.0f} years"
        elif seconds < 31536000 * 1_000_000:
            years = seconds / 31536000
            return f"{years/1000:.0f} thousand years"
        elif seconds < 31536000 * 1_000_000_000:
            years = seconds / 31536000
            return f"{years/1_000_000:.0f} million years"
        else:
            years = seconds / 31536000
            return f"{years/1_000_000_000:.0f} billion years"
