pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ciphersim"
version = "1.0.0"
description = "Educational password security simulation tool for cybersecurity training"
authors = [{name = "CipherSim Team"}]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
keywords = ["cybersecurity", "password", "security", "education", "hashing", "simulation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Education",
]

dependencies = [
    "argon2-cffi>=23.1.0",
    "cryptography>=42.0.0",
    "bcrypt>=4.1.2",
    "rich>=13.7.0",
    "click>=8.1.7",
    "pyyaml>=6.0.1",
    "plotly>=5.18.0",
    "numpy>=1.26.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[project.scripts]
ciphersim = "ciphersim.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
